Implement size & layout checks for the glue layer
This implements a partially-automatic way of
ensuring the size & layout of the structs on the
D and C++ size match.

Note: some types cannot be checked yet due to
the inability to access C++ private fields' offset
information (for example std::vector).

An attempt was made to use friend declarations
instead of making fields public on the C++ side.
This doesn't work as we end up with circular
includes which break compilation in various ways.

Fixes #4