Remove node registry / node set, fix tests with LDC
The intent of `NodeRegistry` / `INodeSet` is very much the same as `Network`:
it provides an interface to manage and communicate with peers.

While communication is technically handled by the `API` object,
some additional features are needed on top, for example the ability
to map a public key to a peer.

This effectively means that 'network' communication in unittests
is now done via `LocalRest`, not via local network anymore.

Additionally, `makeTestNetwork` should be reworked to connect to
the nodes provided in the `Config` for a better API.