Finish discovery when minimum listeners reached
Also add a min_listeners config option
specifying the minimum number of listeners
to connect to before discovery is reached.

Previously, the node's discovery process
would be completed if the number of
connections in progress reached 0.

However this suffers from timing issues.

The discovery could be marked complete even if
the connection and handshake to all the peers
hasn't been established yet.

To simplify discovery and make it predictable,
a min_listeners config option has been introduced.

The discovery of a node will be marked complete
when the number of handshaked nodes reaches
the min_listeners configured value.

Also rename allPeersConnected() to
minPeersConnected(), to better reflect
the use of this function.