Implement the TaskManager class
The TaskManager allows spawning new
tasks that are by default backed
by vibe.d's event loop.

Inherited classes may override this
behavior to spawn tasks which run
in their own defined event loops.

This simplifies code by not requiring
an instance of the Networking class
in order to manage new / existing tasks.

The TaskManager class is subclassed in the
unittests in order to use LocalRest's
event loop, and avoid I/O in the process.