Implement a simple ban manager.
The Ban manager does not interact with any network I/O,
but only simply keeps track of the fail request
stats and marks certain IPs as banned until a
future point in time.

Each time a request to an address fails,
the client code should call onFailedRequest(address),
to keep track of the fail statistics of an address.

The ban manager automatically marks addresses as
banned if they reach a certain number of failed requests.

The client code can call isBanned(address) to check
the ban status of an address.

Once the threshold is reached, the address is banned
for the specified amount of time.

An address may also manually be banned via the ban() API.