Get rid of callbacks and rename getReady()
The use of callbacks here didn't really make sense,
as the function blocked the same task that called
the getReady().

Additionally, the retrieval of the client's peers
depends on minPeersConnected(), which is a property
of the NetworkManager - so the code belongs there.

The getReady() function was renamed to handshake().

Currently it's blocking (no change in semantics),
later it might throw an exception if a number of
failed attempts were made to handshake with the node.