If node fails handshake / netinfo, retry it later
This makes sure that nodes which were previously banned
during the handshake / peer discovery phase get retried
at a later point (at the point they're unbanned),
as long as the peer limit hasn't been reached.

The task that keeps connecting to nodes is now constant
and never stops, as the minimum connected peers may
drop if network connections are lost.