Reduce the number of txs created in a test
Prior to the introduction of multi-tx blocks,
the makeChainedTransactions would create N
transactions.

The API has been changed to create N * TxsInBlock
amount of transactions, but the test-case still
used the same N.

Reduce the number of tx created to avoid spawning
too many fibers when propagating transactions.

In the future, the transaction propagation,
and the issuance of new fibers, should be handled
by a fiber manager.