Split the API management out of TestNetworkManager
The class was being instantiated not only per-node,
but also one additional instance per-unittest.

The unittest instance however was not really a proper
NetworkManager instance, as the majority of the fields
inside of NetworkManager were left null and would
cause crashes if any method on it was called.

This commit introduces a TestAPIManager, whose sole
role is to manage the starting / stopping of nodes,
and issuing requests for each node in the unittest.