Attempt retrieving blocks from multiple nodes
Previously, the node retrieved blocks from the node
which had the highest height.

However this opens the possibility for a DoS,
as a node can simply lie about having the highest
height.

Instead, try downloading from all nodes until a
valid (!) block with the highest block height is
retreived from the network.