Convert system integration test from shell script to D
It does not look like a win right now,
but it's going to look much better once we add more logic.
One of the advantage we're already earning from this is that
building the image is not dependent on the working directory anymore,
meaning this script can be called from anywhere, not just the root directory.