Correctly parse the Quorum config section
The Quorum is made out of one or more nodes,
and one or more sub-quorums, up to a number
of nesting levels.

The 'name' option for a quorum has been removed,
as now the configuration is simplified by
using nesting rather than backreferences.

The threshold has been changed to an N of M integer,
which will make it easier to interface with SCP
which uses the same methodology.

Fixes #265