Improve and correct PrettyPrinter output
The previous output was not correct, and could have one zero too much.
In addition, it was too verbose for monetary values.
The intent was to make monetary value formatted to a constant size,
but since it is not done for the integral part, it's useless to do it
for the decimal part.
Making the integral part constant-sized is also not an option.