Make starting the nodes asynchronous()
Currently there isn't a simple way to ignore
return messages from LocalRest calls, therefore
the calling thread gets stalled waiting for an
answer before it runs start() on all the nodes.

As a temporary workaround, we can use threads
to issue the start() call on each of the nodes.

The waiting time between handshake retries in the
test-suite has been reduced, as the first call to
getDiscoveredNodes() will be called before the
nodes have all handshaked.