Make coverage work with integration tests
The only thing we need to do is to mount `source` and `submodules`
in the working directory so that the coverage code can access it
(otherwise it's unable to generate reports).
Since coverage won't be enabled very often, or by end user,
the option to store it in the data directory is removed
(it also became moot since the reshuffling of the docker image).