Use workaround for caling std::function delegates
Previously, the SCP C++ code allocated an std::function
which would be collected at the end of the function scope
because the reference count remained at 1.

To work around this, the delegate is now assigned to
a pointer allocated on the heap.

Once the delegate is invoked, it's then destroyed
as it will only be called once.

The setupTimer() API was changed to take a pointer to
an std::function, rather than pass it by value.

The callCPPDelegate() wrapper was introduced,
which calls the provided std::function*,
and deletes the pointer after the function call.