Add validators and signature to the BlockHeader
The BlockHeader now contains:

- A storage for the bitfield of validators which signed the block
- A Schnorr based multisignature field that contains the
  signatures of all the validators which signed the block header.

The bitfield is an array of bits, where the index into the
array is the Nth validator in the known set of validators
which previously enrolled and are still enrolled.

The set of known validators must be sorted for this to work.

Example usage (pseudo-code):

------
Point sum_keys =
    block_header.
        .bitfield  // range over bits
        .filter!((idx, has_signed) => has_signed))  // validators which signed
        .map!((idx, _) => this.validator_keys[idx])  // get the keys only
        .reduce!((a, b) => a + b));  // sum them via schnorr '+' operator

// verification
if (verify(sum_keys, block.signature, block.hash))
{
   // ...
}
-----