Remove hashing requirement to have `computeHash` implemented
`hashPart` used to require that `computeHash` be implemented by aggregates.
However, this proved very inefficient and led to both abstraction leaks,
and repetitive code.

This commit lifts this limitation, using 'tupleof' as the default hashing mechanism.
Aggregates still have the ability to override the way hashing is done
by explicitly implementing `computeHash`.

Note that not having to implement `computeHash` makes aggregates with special requirements
(namely `Block` and `Input`) more obvious to the reader,
as otherwise the override gets lost in the noise.