Travis: Use different dub on DMD 2.089.1 + Linux
It seems that currently DUB suffers from a druntime regression,
so this should help with the random segfaults in the CI.
After investigation, it might be originating from the parallel
marking done by the GC.

However, due to a discrepancy in 'install.sh', the 'dub' that can
be directly installed is compiled with an older compiler than the
one bundled with DMD 2.089.1, and this older compiler does not
support parallel marking.

However, this older compiler also does not have the correct code
to support TLS data on Mac OSX 10.15, hence why the matrix is split.
The parallel marking bug does not seem to manifest on OSX.