Remove ConfigException, use Exception directly
ConfigException was an attempt at simplifying the code that didn't turn quite right.
The whole point of having it seemed to avoid declaring the 'Config' variable outside
of the 'try' block, but it resulted in an extra level of indentation for everything,
and a try-catch-throw which worked around the fact that only the YAML parser throws.