Implement validation routine for a Block signature
Using Schnorr and the preimage revealing scheme,
the new isInvalidSignatureReason() / isInvalidSignature()
validates the block signature.

The routines take the block header, a map of the
revealed preimages, a map of the R values which were
calculated through the use of preimages,
and the array of sorted public keys of all the validators
which have been enrolled in previously finalized blocks.

The signing of a block should proceed as follows:

- The nodes signal whether they will sign the block.
- The sum of all the nodes' P and R which signaled
  they will sign the block becomes the P and R used
  to sign & validate the block.

The validation of a block's signature has a number of checks:

- The validators array (bitfield of nodes which signed)
  must not exceed the number of enrolled validators.

- The preimages for the block at this height must have
  been revealed (provided by the preimages_map parameter)

- The R's for the block at this height must have been
  calculated based on the preimages which were revealed.

- The R in the Schnorr signature must match the expected
  R based on which validators signed (bitfield array).

- Finally, the Schnorr signature itself is validated.

Note: Each validating node must also make sure to remove
another node's key from the provided 'pub_keys' if that
node's validation cycle has ended.