System test: Correct path to build.d
The path to build.d was incorrectly referencing 'source' while it should have been
'../../source/'. This was probably working because of a bug which has since been fixed.
The CI didn't pick it up for so long because the package was most likely cached
or still present in the testing repository, but it started to fail recently.