Fix the timer handling in setupTimer()
When SCP calls setupTimer() with a timeout of zero,
it wants to cancel all previously started timers.

However, there is a potential timing issue. Previously we added and
removed timers based on the 'timerIDs', which are not actually
IDs but are types of timers (there are only two types: ballot and nominator).

A possible scenario (pseudocode):

- setupTimer(1.seconds, ID.Ballot, { firstVote(); });
- ... 500 msecs later
- setupTimer(0.seconds, ID.Ballot, null);
  => this "deactivates the timer"
- ... 100 msecs later
- setupTimer(1.seconds, ID.Ballot, { secondVote(); }); // new timer
- ... 400 msecs later
  => the first timer now still gets invoked, even though it was "cancelled"

The reason for this bug is because we used a single hashmap for all timers
based on the timer type.

This has been changed, so now all timers have a unique incrementing ID,
and when a timer of a specific type is cancelled, then all *current* timers
will be cancelled, but any new ones will continue to work.

Creating a proper test-case is difficult due to this being a C++ method
that's invoked by SCP and uses C++ callbacks and the task manager.