Move nomination responsibility to Nominator class
The Ledger class suffered from leaky abstractions,
acting as both a ledger-tracking class and as a
nominating class, conflating two different
responsibilities.

The previous several commits have gradually shifted
that responsibility to the Nominator class.

This will also make it easier to change the nomination
protocol to be timer-based, rather than trigger-based
as it is now.

Fixes #572