Serializer: Always do serialization of uint[] elements (and other)
There was previously a mismatch in the way uints were handled,
leading to inconsistent behavior depending on whether or not
the uint was an array, or in a struct, or an array in a struct.

Here we remove the special case for uint[] in the serializer,
and most arrays but `ubyte[]` and `[{w,d}]string` in the deserialized,
so that the results are consistent.

It might have an impact on (de)serialization's performance,
however it is currently not an issue.