Make MemBlockStorage behave more like BlockStorage
After BlockStorage is initialized, it is not yet usable,
as load needs to be calledd.

However, MemBlockStorage is much more forgiving as it is
fully initialized after the constructor call,
and calling `load` is a no-op.

This is one aspect of dependency injection which is tedious:
the injected type needs to behave as closely as possible
to the original type in order for tests to be meaningful.