Fold all serialization routines into one
Originally, both deserialization and serializer had multiple overloads
that each handled a specific type.
However, as the types being covered expanded, and usage of serialization grew
(it was originally intended to serialize types on which we had full control,
e.g. blocks, but will now be used for inter-node communication),
this approache did not scale: to provide any kind of generic overload,
e.g. for arrays, it is necessary to exclude the special case handled by
other overloads, for example the `string` or `ubyte[]` specialization.
The approach of merging overloads was already taken in the Deserializer.