Merge Deserializer into Serializer
The separation between the two was an illusion, as can be illustrated from the changes:
not a single module imported Deserializer without importing Serializer.
This merge makes usage easier, as well as unittest code,
which previously was comparing binary data with a serialized or deserialized value.
Some places still need to compare binary data (e.g. to ensure we don't encode a length for static arrays),
other are deserializer-specific tests (e.g. when fed wrong input), and are marked as such.
The documentation was greatly extended, and documents some upcoming improvements,
such as the static 'fromBinary' method which will be part of a later PR.