Remove boolean status code from receiveEnvelope()
Clients should not expect any status code when issuing calls
to the receiveEnvelope() API.

The node itself may care whether its internal SCP class
considers the envelope valid or not, but it does not need
to signal back to the client whether it accepted the envelope.

This is because this API should be used in an asynchronous way.
Aftera client floods a given SCPEnvelope to the network,
it does not care about how that envelope is further processed
by each node.