Implement a getAddress() abstraction
The NetworkManager in the unittests uses the public
key of the node as the effective address,
but then the base class also used IP-based address
banning despite of that.

By abstracting the address, we can keep the behavior
predictable.

In addition, we will need to pass the address
later when the handshake stage is changed later
in preparation to fix #606