Fix undefined identifier error on v0.x.x
This is currently triggering because two PRs that conflicted were rebased.
In particular, the logging was extended, but tx_set was wrapped in ConsensusData.
This also invert the order of argument, putting the reason first,
as it makes it more visible, both by being first and by being much shorter
than ConsensusData.