Make Node.start() always run in a task
This allows the test-suite to start multiple nodes
in parallel without waiting for the discovery phase
of one node to complete before another node's discovery
begins.

Without this we cannot test network discovery properly
due to a race condition:

A.start() => calls B.getNetworkInfo()

However B's network info will always be empty because B.start()
was never called, because the main thread is stuck waiting
for A.start() to return.

Fixes #276