Change the mechanism of creating preimages
Previously, we create only 1008 number of pre-images from a random value.
But there is no way to restore pre-images in abnormal situations like
database crashes. So, we have decided to change the mechnism of creating
pre-images. we create a great number of pre-images like a hundred of
thousand from the private key of a node.
If re-creation of pre-images is needed, the enrollment manager creates
new bulk of pre-images with a salt hash. A salt hash is derived from
a bulk index which is an order of a series of a bulk of pre-images like
0, 1, 2, ... N.