Remove ForwardCtors from opaque_array (and scpd.Util)
In order for deserialization to support those types,
there must be a constructor that accept a `tupleof` the struct.
Hence we can replace `ForwardCtors` with an explicit constructor.

After removing the last `ForwardCtors` usage, we can remove it
alongside the supporting code.
This leaves the `scpd.Util` module with nothing but one extern declaration,
which clearly doesn't pass the bar for a single module,
so the declaration was moved in the most appropriate module, `scpd.Cpp`.