Drop support for DMD
DMD backend has proven to be a constant source of issues.
It produces suboptimal code, and behaves differently from other compilers,
e.g. causing unwarranted and unpreventable moves which would be detrimental
to our usage of `std::string`.
Last but not least, it provides completely broken debug infos on OSX.