Add ability to provide a custom Genesis block
GenesisBlock becomes a getter function which
returns the genesis block.

The genesis block is by default initialized to the
CoinNet Genesis block in non-unittest builds,
and to a UnitTest Genesis block in unittest mode.

There is a new config option in the Node config
which allows overriding the default genesis block,
by setting the 'genesis_block' config option to
the hexadecimal dump of a serialized genesis block.

This allows system integration tests to use their
own genesis block while testing the non-unittest
build of Agora.

Additionally it allows easy deployment of TestNet
or any private blockchain networksa user wishes
to use.

If a unittest wishes to test with a custom genesis
block, they can do this:

```D
unittest
{
    immutable old_gen = GenesisBlock;
    scope (exit) GenesisBlock = old_gen;

    setGenesisBlock(new immutable(Block){ ... });
    // ...
}
```

Fixes #709