Merge push and pr, test on Windows, and test all compilers
Currently we hard fail on PR and pushes the don't pass with {dmd,ldc}-master.
This can be disabled by uncommenting the `continue-on-error` statement.
However, since we hadn't have an issue for a while, let's start with this.
This revert some of the previous changes that were done to simplify the CI,
but we hope that reducing the dimensions (in the future) will make up for it.