Add SCP's quorum intersection checker
See also: https://github.com/stellar/stellar-core/pull/2127

Also made changes to make the Quorum checker buildable

In stellar-core v11.2.0 the quorum intersection checker
depends on Stellar's Config struct.

This dependency was removed here so we can use the code.
Additionally shared_ptr was removed from one of the
parameters as it was causing linker issues,
plus we do not have a complete binding to shared_ptr
with the correct semantics yet.

Co-authored-by: MonsieurNicolas <nicolas@stellar.org>