SCP: Disable very chatty TRACE messages
Those messages are triggered very often, and of little use to us.
While TRACE output should be somewhat liberal in what to output,
having side-effect-less functions being so verbose tends to make
the TRACE output much less readable.
After this change, most, if not all, of the messages relate to Ballot.