Rename updateEnrolledHeight to addValidator
As we are splitting the EnrollmentManager into EnrollmentPool and ValidatorSet,
this function had to be renamed to fit its new purpose.
Before, enrollment that were not in the chain would have an enrolled height of 0,
and when they appeared in the chain, the actual enrolled height would be set
through this function.
Soon the storage will be separate, and instead this function will just add new enrollments
and clear up the pool.
Additionally, this fixes a bug, where catchup didn't work when an Enrollment seen in a block
was not found in the Enrollment pool.