Split enrollments into candidates and validators
The ValidatorSet had an issue that it manages two kinds of enrollments
which consist of the enrollments waiting to be a validator and the
enrollments having already been a validator. This could cause an
unexpected error, so we split the class into the EnrollmentPool for
the candidates of a validator and the ValidatorSet for the confirmed
validators.