Restore a node's enrollment data from the header of block
This code restores the enrollment data of a node from the `enrollments` field
of the header of blocks. This commit is the first step for remove the
`node_enroll_data` table, which is to remove the `Enrollment` object from the
table. The next step is to remove `next_reveal_height` from the table. Finally
the table will be removed.