Add validatorCount() to EnrollmentManager
This is a less costly operation than calling
getValidators(). Client code may want to
call validatorCount() twice between calls
to clearExpiredValidators(), thereby easily
determining if the set of validators has
changed, and act accordingly.

For example, if a Validator's cycle has expired,
the Quorum configuration will need to be regenerated,
as described in #785