Clear genesis enrollments properly
The ValidatorSet assumes that enrollments can only
be enrolled starting from block height #1.

However, the genesis block can (and must) include
a set of validators.

Validators always sign blocks starting from
enroll_height + 1. For example, a validator which is
in the genesis block will sign blocks [1 .. 1008].

When block 1008 is externalized, this validator
should be removed from the validator set.