Correctly reset the cycle counters & populate the pre-image cache
The previous code was buggy, as it would take the last enrollment (first in the array)
instead of the first enrollment (last in the array) for pre-image propagation.
In addition, there was an off-by-one error in the reset, which means that the last
`consume` of the cycle would not correctly work.