ValidatorSet tests: Use correct keys and utxo hashes
In order to make the test more DRY, and accomodate for future improvements,
the transactions are now sent to 5 different keys (4 are used in the tests),
whereas before all funds were controlled by the same node, which enlisted multiple times.
Additionally, this uncovered a bug with the test, as the return value of '.keys'
is not sorted, hence the utxo hash AA was changed to a simple array,
and much duplication was removed.