tests: Increase default timeout to 5 seconds
We did not account for the fact that our unittests
are built with code coverage support.

For each line in the D code, the compiler injects
an increment opcode do a compiler-built bit array.

This can significantly slow down processing,
so as a counter-measure we can increase the
timeouts to make the test-suite less brittle.

The ultimate solution would be to replace the use
of real time with virtual time. See #389.