Github CI: Split artifact by matrix dimension
A problem that was witnessed was that artifacts were all mashed together,
because they all used the same name.
Solve this problem by using os and dc as part of the name.
However, for push and releases, we don't want to upload artifacts for all 3 compilers,
instead we want to select the compiler that is used for official build.