The `EnrollmentManager` add enrollments before the cycle
The `EnrollmentManager` should allow a node to re-enroll before the validator cycle
ends. If requests for enrolllment arrives before the cycle ends, the
enrollment will be added in the height where the cycle ends. So the
`getEnrollments` of the `EnrollmentManager` must return the enrollments right
before the end of the cycle.

Relates #824