Rework NetworkManager by using tasks
The new NetworkManager now uses separate connection
and address discovery tasks.

The 'connection_tasks' config option has been added
to fine-tune the number of connection tasks a Node
should use.

The discover() method can now be called multiple times,
which is required to be able to support quorum reshuffling.

The address discovery is now an endless task which does
not need to be re-triggered, it runs in the background and
queries each client in turn every 5 seconds.

The periodic catchup routine no longer waits for full
node discovery before attempting to read blocks.
It is unnecessary to have to wait for minPeersConnected()
to be true, catchup can proceed as soon as we have
at least one node connection.

Fixes #587