Don't ban our own IP, filter explicitly
Using the ban manager with an infinite
ban time is awkward, and further complicates
the situation where the node's IP changes.

The Ban data is persistent and this is problematic
if the Node's IP changes but we still have an
IP in the ban list that is no longer used.