Move UTXOSet to agora.consensus from data
The data module is used for data types which are exchanged to achieve consensus.
In essence, they contain the operations that are applied to the state machine.
On the other hand, the UTXOSet is the state itself, the result of applying the operations.
Hence, it always stood out as an outlier in that package.
Moving it in the same package as ValidatorSet,
a module it shares a lot of ideological similarities with, makes much more sense.