Implement shell quorum balancing
The quorum set is now derived from the active
set of enrollments in the blockchain.

After each block has been externalized,
the active validator set is checked.
If the set of validators has changed,
then the quorum set for any Validators
is reconfigured according to the
quorum generator algorithm.

Fixes #785