Add ability to track double spends in `TestUTXOSet`
This should bring it more in line with the actual UTXOSet.
However, since many tests do not care about double-spend prevention,
and tests tend to use a more limited subset of transactions,
we keep `peekUTXO` exposed.
Also, as many tests currently depend on `findUTXO` being `peekUTXO`,
an alias was added to avoid an avalanche of changes.