Create a class managing consensus constants
In order to set the consensus-critical constants in runtime and share them
between every object, we must provide only one parameter set. So we worked
on this with the concept of the Singleton pattern in order to meet the
requirements.

The only `Params` object exists after being created with the `createParams`
function that calls the constructor of the class.