Use atomic database transactions
The Ledger will now issue a begin statement
to all managed SQLite databases before
any changes are made in the call to
addValidatedBlock().

Upon success, commit will be issued
and all of the changes will be stored to the
managed databases.

If there was a failure (exception thrown),
rollback will be issued. The rollback will
cancel any pending transactions started
after commit.

If the node crashes before commit is issued,
all of the pending transactions will be
lost and therefore rollback is unnecessary
(and not possible since it was a crash).

Part of: #836