Add `ConsensusParams` class for consensus-critical constants
For unit-tests, it's not efficient to test with the constants for consensus
like the `ValidatorCycle` because the original value make the tests slower and
very time-consuming. The goal of unit tests is checking the operations run
properly in test environment.

In this commit, we provide the definition for configurable consensus-critical
constants, or the `ConsensusParams`.