Don't throw on getMerklePath() failure
A failure to read a block from the persistent
block storage should not cause the Node
to throw exceptions in the call to getMerklePath().

Furthermore we must not allow external clients
to know a request failed because of an
internal thrown exception as this
could become a DDoS attack vector.

Also mark many methods as 'nothrow'.