getBlocksFrom() should return mutable array
There is no point in returning a const array,
because the caller is the sole owner of this
memory. It alone can decide whether it wants
to pass this data to a function which promises
not to mutate the data.

This also avoids the need to do deep copies,
such as in the previous call to:
this.last_block = ledger.getLastBlock();

In this case what we were trying to achieve
was a deep-duplicate of a Block just in order
to have a mutable block stored in the Ledger
and used as a cache.