Improve Ledger testability and reorder constructors arguments
Re-order the constructor arguments so that it is more sensible.
The most important arguments are the one the ledger cannot do without,
namely the configuration, consensus parameter, and various utilities
to handle consensus rules.
Things like TransactionPool could actually be optional
(even though they currently aren't).

Secondly, the Ledger unittests are actually very repetitive,
and there is little need to change e.g. the TransactionPool used,
or the `EnrollmentManager`. So we provide a unittest-only subclass
that has all the sensible defaults.