Set the limit on the validator cycle
We have made the consensus-critical constants set through a parameter in the
last sprint. By the way, developers could set the validator cycle to a value
smaller than the RevealPreimagePeriod value inadvertently, which make the
developers not to be able to find the cause of failures about tests of
revealing pre-images.

So I set the restriction that the value must be equal to or greater than the
validator cycle.