Fix memory corruption in combineCandidates()
It was returning pointers to GC-allocated memory
to C++, however the GC would then reclaim this memory
as all the references to it were only held in SCPP.

Later when the Ballot timer would fire, it would
re-elect a Value but the validation would fail
as the memory it held failed to be deserialized
(it was reused).

Instead just return a duplicate of a vector,
which will later be reclaimed by its destructor.