Implement stake-based quorum balancer
Nodes which have a higher stake compared
to the rest of the network will have a
higher chance of being included in other
node's quorum sets.

There are two constants which alter the
generator's behavior:

- MAX_NODES_IN_QUORUM
  Maximum number of nodes to include in a quorum.

- THRESHOLD
  Threshold to use for the quorum configuration (e.g. 0.80 = 80%).

The algorithm uses a random number generator to shuffle
the quorum node selection. The public key of a node is used
to initialize the RNG. In a later edition we will also
combine it with the concatenation of all preimages to
make the shuffling less predictable.

Fixes: #963