Trivial: Grep for the error message in validation tests
Validation tests are a bit finicky, as they return a string literal
to describe the problem.
Comparing to the string literal would create unnecessary coupling,
but just checking that the test fails can easily lead to false positive.
Grepping for a key component of the message is a middle ground.