Genesis: Reduce UnittestGenesisBlock assumption on being single-tx
In many places, it was assumed that the genesis block was single-transactions.
Two examples are the usages in `makeChainedTransactions` and the merkle root
being used as merkle tree.
To reduce the diff that will arise when we move from one transaction to two
(a payment and a freeze one), this replace the merkle root assumption,
and remove the 'genesis transaction'.