Allow to enroll at the same time as freezing
This is required for Genesis to have Enrollments.
Instead of specializing it for Genesis, this is currently done for all blocks,
as there is no obvious downside and it seems desireable considering current work
(namely not accepting blocks that would render the chain inoperable).

The implementation is however quite simple and needs to be cleaned up in the future,
as it uses as rather simple test utility for this.