Add `avail_height` column into `EnrollmentPool`
We add the `avail_height` column into the `EnrollmentPool` in order to check
if an enrollment can be a validator. If even the enrollment had been promoted
as a validator, it can be used as the request to be a validator in the next
cycle. So the `getEnrollments` function can select the enrollments which can
be a validator for next block with the `avail_height` and the height passed
in as a parameter.