EnrollmentManager: Remove buggy call to setEnrollmentKey
This call is obviously wrong: it uses 'enroll' as a parameter,
which at this point is default-initialized.
So if an enrollment exists, the update would overwrite valid data
with all default-initialized data (most likely all 0).
After removing this wrong call, the 'setEnrollmentKey' function
is left unused, and thus was removed, without any effect on the tests.