Take NodeConfig.validator_cycle into account
Before this change, this config option was simply ignored,
be it in the config file or in the unittests.
Instead, the unittests would pass a `ConsensusParams` instance directly.
Changing this to have the node read it from the config simplifies
code a bit, and allow for future enhancement.
It also removes the need to add more consistency checks.