Use preimages for the random seed
The getRandomSeed() function retrieves all
the active validator's random seeds for
the given block height.

A validator uses this random seed in
the quorum generator.

The onValidatorsChanged() callback has been changed
to take the block height rather than the implicit
latest block height, as during boot we only have the
commitments in the genesis block and *not* the preimages
for the latest block height.

Currently these preimages are only used for the quorum
generator - and the quorums only change when there are
new enrollments (and therefore new commitments).

In the future the quorums will be regularly reshuffled
and we will need to be able to restore the preimage information
when the node boots up (currently missing)

Fixes #990