Implement getRandomSeed()
This EnrollmentManager function generates the
random seed which should be used to feed the
quorum generator with.

It takes the set of enrolled UTXOs and the
desired block height.

If the block height is in the past compared
to the enrolled height + distance of a preimage,
the preimage will be hashed to get the correct
preimage of the desired block height.

getRandomSeed() then concatenates the hash
of all the preimages for the given height,
and returns the hash to be used as a random seed.