Work around hashmap invalidation bug
Walking over a hashmap and doing a fiber context
switch means that another fiber (e.g. DiscoveryFiber)
could modify the hashmap and rehash it.

This would invalidate the iterator of the first
fiber and can cause a segfault later when the
fiber wakes up and tries to dereference a pointer to
reclaimed memory.

As a temporary workaround, simply use an ever-growing array.