Fix race condition when spawning connection tasks
The constructor of ConnectionTask previously
started the task immediately, which meant
that it doesn't always have the chance of
being stored in the connection hashmap.

Change the code so it stores the connection
task in the hashmap, and then start the fiber.

Once the connection task is complete it will
remove itself from the hashmap.