Restore a node's enrollment data in adding validator
The `addValidator` function is the best place to restore a node's enrollment
information like the key for enrollment and the pre-images.

We introduce a new function named `populate` in the`PreImageCycle` class that
is replacing the `consume` function. The new function have the option to
either just get the first pre-image or calculate the cycle index.