Use the test GenesisBlock in the integration tests
This allows things to be much more predictable, e.g. by allowing us to use well-known keys.
It also brings it back in line with the rest of the code and hopefully will allow
us to remove the 'genesis_block' string in the config file after we split the genesis block module.

A few changes are required as a result:
- Node count goes from 3 validator, 1 full node to 6 validator, 1 full node;
- Private keys of nodes are changed to that of NODE2 .. NODE7;
- Nodes were re-shuffled to matche their public address:
  - Ports were changed: NODE2 has 2826, NODE3 has 3836, ..., NODE7 has 7876;
  - Folders were also renamed, so node/2/config.yaml contains the seed for NODE2;
  - The full node is in folder 0 and port 1826;

In addition, the client didn't know what is the "correct" genesis block at the moment.
It just used its own Genesis block, which was arguably a bug.
It will now query the genesis block from the nodes and use that to make transactions.