Save the cycle index of pre-images
When the `addValidator` is called, the cycle count for enrollments should
be restored. First, we add new column named 'cycle_count` into the table, and
when the `addValidator` function is called, the `cycle_count` is saved. After
that, when the `EnrollmentManager` restarts, it loads the count and calls the
`populate` function as times as the count.